# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fff-server",
    "description": "Typed HTTP access to the fff file-search engine.",
    "license": {
      "name": "Unlicense",
      "identifier": "Unlicense"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default local bind"
    }
  ],
  "paths": {
...
    "/v1/workspaces/{id}/glob": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Literal glob match, frecency-ranked, bypassing the query parser.",
        "operationId": "glob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Matching files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid glob pattern",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Index not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{id}/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Fuzzy file search.",
        "operationId": "search",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ranked file matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Index not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{id}/search/directories": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Fuzzy directory search.",
        "operationId": "search_directories",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ranked directory matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirSearchResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Index not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{id}/search/mixed": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Fuzzy search over files and directories together.",
        "operationId": "search_mixed",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ranked file and directory matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MixedSearchResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Index not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirHit": {
        "type": "object",
        "required": [
          "item",
          "score"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/DirItemDto"
          },
          "score": {
            "$ref": "#/components/schemas/ScoreDto"
          }
        }
      },
...
      "DirSearchResponse": {
        "type": "object",
        "required": [
          "items",
          "totalMatched",
          "totalDirs",
          "page",
          "pageSize",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirHit"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "minimum": 0
          },
          "totalDirs": {
            "type": "integer",
            "minimum": 0
          },
          "totalMatched": {
            "type": "integer",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "GitStatusConstraint": {
        "type": "string",
        "description": "Git status filter, mirroring the engine's four categories.",
        "enum": [
          "modified",
          "untracked",
          "staged",
          "unmodified"
        ]
      },
      "GlobRequest": {
        "type": "object",
        "description": "Glob-only request. Bypasses the query parser entirely: the pattern is used literally,\nresults are frecency-ranked. Use it when the pattern is already a glob and fuzzy\nmatching on top would only add noise.",
        "required": [
          "pattern"
        ],
        "properties": {
          "currentFile": {
            "type": [
              "string",
              "null"
            ]
          },
          "maxThreads": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "page": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "pageSize": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "pattern": {
            "type": "string",
            "description": "Forward slashes only — the matcher does not treat `\\` as a separator.",
            "example": "**/*.rs"
          }
        },
        "additionalProperties": false
      },
...
      "LocationDto": {
        "type": "object",
        "description": "A `file.ts:42:10` suffix parsed out of the query. `line` is always present; `col`,\n`endLine` and `endCol` depend on `type`.",
        "required": [
          "type",
          "line"
        ],
        "properties": {
          "col": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Present for `position` and `range`."
          },
          "endCol": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Present for `range` only."
          },
          "endLine": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Present for `range` only."
          },
          "line": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/LocationType"
          }
        }
      },
      "LocationType": {
        "type": "string",
        "description": "Which parts of a location the query specified.",
        "enum": [
          "line",
          "position",
          "range"
        ]
      },
      "MixedHit": {
        "type": "object",
        "description": "A hit from a mixed search. Exactly one of `file` and `directory` is populated, indicated\nby `type`.",
        "required": [
          "type",
          "score"
        ],
        "properties": {
          "directory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DirItemDto"
              }
            ]
          },
          "file": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileItemDto"
              }
            ]
          },
          "score": {
            "$ref": "#/components/schemas/ScoreDto"
          },
          "type": {
            "$ref": "#/components/schemas/MixedKind"
          }
        }
      },
...
      "MixedSearchResponse": {
        "type": "object",
        "required": [
          "items",
          "totalMatched",
          "totalFiles",
          "totalDirs",
          "page",
          "pageSize",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MixedHit"
            }
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "minimum": 0
          },
          "totalDirs": {
            "type": "integer",
            "minimum": 0
          },
          "totalFiles": {
            "type": "integer",
            "minimum": 0
          },
          "totalMatched": {
            "type": "integer",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. ASP.NET Core deserialises this into `ProblemDetails` with no\nclient-side code.\n\nClients switch on `code`, never on `detail`: `code` is a stable identifier, `detail` is\nprose from the engine and may change between fff releases.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code. Switch on this.",
            "example": "invalid-path"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Prose explanation for this occurrence. Not stable; do not parse."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated here as RFC 9457 specifies.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the error class."
          },
          "type": {
            "type": "string",
            "description": "Stable URN identifying the error class.",
            "example": "urn:fff-server:error:invalid-path"
          }
        }
      },
...
      "SearchHit": {
        "type": "object",
        "description": "One file hit, with its score and match highlighting.\n\nThe engine returns items, scores and match offsets as three parallel arrays; they are\nzipped here. Nothing is lost, and a generated C# client is far better for it.",
        "required": [
          "item",
          "score",
          "matchRangesUtf16"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/FileItemDto"
          },
          "matchRangesUtf16": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Utf16Range"
            },
            "description": "Which parts of `item.relativePath` matched, in UTF-16 code units. The MCP server\ndiscards this information entirely."
          },
          "score": {
            "$ref": "#/components/schemas/ScoreDto"
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "description": "Search request. Supply `query`, or `structured`, or neither (an empty query ranks the\nwhole index by frecency). Supplying both is rejected rather than silently preferring one.",
        "properties": {
          "comboBoostMultiplier": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Query-history combo boost. Only has an effect once a client reports selections via\n`track-query`."
          },
          "currentFile": {
            "type": [
              "string",
              "null"
            ],
            "description": "A path to deprioritise, so the file a caller is already looking at does not dominate\nits own results. Also feeds the distance bonus for nearby paths."
          },
          "maxThreads": {
            "type": [
              "integer",
              "null"
            ],
            "description": "0 lets the engine choose.",
            "minimum": 0
          },
          "minComboCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "page": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Zero-based page index.",
            "minimum": 0
          },
          "pageSize": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Defaults to `defaults.page_size`.",
            "minimum": 0
          },
          "query": {
            "type": [
              "string",
              "null"
            ],
            "description": "The raw fff query DSL, e.g. `git:modified src/**/*.rs !tests/ user controller`.\nPassed to the engine byte-for-byte. Glob separators must be `/`.",
            "example": "src/**/*.rs button"
          },
          "structured": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StructuredQuery",
                "description": "Constraints as data, instead of a DSL string."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "SearchResponse": {
        "type": "object",
        "required": [
          "items",
          "totalMatched",
          "totalFiles",
          "page",
          "pageSize",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            }
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LocationDto",
                "description": "A `file.ts:42:10` suffix parsed out of the query, if present."
              }
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "minimum": 0
          },
          "totalFiles": {
            "type": "integer",
            "description": "Files in the index.",
            "minimum": 0
          },
          "totalMatched": {
            "type": "integer",
            "description": "Matches across the whole index, not just this page.",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Non-fatal notes about the query — currently backslashes in glob-looking tokens,\nwhich would otherwise match nothing with no explanation."
          }
        }
      },
      "StructuredQuery": {
        "type": "object",
        "description": "Constraints as data, for clients that would rather not build DSL strings.\n\nSeparators are folded in `globs`, `pathSegments`, `filePaths` and `excludePathSegments`,\nwhere a separator is unambiguously a separator. The engine's glob backend treats **only**\n`/` as a separator, so without folding `src\\**\\*.rs` would not be applied as a glob at\nall — it would fall through to fuzzy text matching and quietly return unrelated results.",
        "properties": {
          "excludePathSegments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Path segments to exclude."
          },
          "excludeText": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Literal text to exclude."
          },
          "extensions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bare extensions, without a dot: `rs`, `md`."
          },
          "filePaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Exact repo-relative file paths."
          },
          "fileTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine file-type names, e.g. `rust`."
          },
          "fuzzy": {
            "type": [
              "string",
              "null"
            ],
            "description": "Free-text fuzzy terms. Split on whitespace into parts, as the DSL parser would."
          },
          "gitStatus": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GitStatusConstraint"
              }
            ]
          },
          "globs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Glob patterns, e.g. `src/**/*.rs`."
          },
          "pathSegments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Path segments a result must contain, e.g. `src`."
          },
          "text": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Literal text a path must contain."
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}